編集ページでの画像Preview機能のhook化
